Rewrite LogToBigQuery as a Dataflow streaming job
Summary:
This is a rewrite of the relevant LogToBigQuery code as a Java Dataflow job.
This gives us real-time logs formatted in exactly the same way as in the `logs`
dataset, including the `elog` fields and bingo events. In the long run, when the
Python Dataflow SDK comes out, we may want to rewrite the code itself in Python,
but this gives us some experience managing Dataflow jobs, and having the
streaming logs sooner is nice.

This is already working in production! You can see the job running here:

https://console.developers.google.com/project/khan-academy/dataflow/job/2015-09-08_02_50_35-12511464375788774586

And you can try it out yourself by querying the `logs_streaming.last_hour` view
in the `khan` project or the `logs_streaming.logs_all_time` table in the
`khan-academy` project.

Here are some of the details of the implementation:
* The app engine logs are now pointed at the `request_logs` topic using the
  "exports" tab of the "logs" section of the developers console.
* A subscription called `log_export_prod` exists for that topic and is used for
  every job to ensure that we don't drop logs or have massive duplication.
* The code itself subscribes to that subscription, transforms (using a simple
  `ParDo`) each log line into a BigQuery row, which is then sent to BigQuery
  using the BigQuery streaming API.
* There were a few places where the `protoPayload` format differs from the
  original App Engine requestlogs API, so I had to add some special cases.
* Most of the Google JSON formats here have auto-generated Java client
  libraries, so I used those where I could.

There are still plenty of loose ends here:
* Tables don't roll over, there's just one table that grows forever. The plus
  side is this makes it easy to implement a `last_hour` view.
* There is no mechanism for automatically making schema changes like there is in
  the Python batch job. That means that you need to manually update the schema
  when adding a field, or blow away the table and start over.
* There's no mechanism to keep the event log fields in sync with webapp.
* Everything runs in the `khan-academy` project, including the output BigQuery
  table. I tried, but I couldn't get the authentication figured out to be able
  to write the output to the `khan` project.
* There are no tests, despite the code being quite testable.

Test Plan:
Use the Makefile to run the job in production. I iterated on a few errors that
showed up and now it has been running for hours without errors. The output
itself seems sensible.

Reviewers: chris, colin

Reviewed By: colin

Subscribers: davidwang, #infrastructure

Differential Revision: https://phabricator.khanacademy.org/D21531