Set up the log export job to be runnable locally
Summary:
While testing out the dataflow streaming log export job (which turned out to
work fine as-is), I set it up so that it's possible to test it locally
(i.e. without sending it to the dataflow service).  This makes the debugging
cycle much, much shorter.  This revision adds these changes so that it's easier
to hack on this dataflow job in the future.

The main changes I made were to the Makefile, which adds a new option
`run_locally` that uses the `DirectPipelineRunner` which is kind of like the
devserver for dataflow and just runs the job on your machine.  I also added a
command line configuration `--runBoundedOver=N` that lets you cleanly run it
over just a fixed number, `N` records locally, since normally it's an infinite
streaming job.  (With this parameter, you still have to Ctrl-C out of the
pipeline, but it stops reading new inputs after `N` records.)

Finally, I changed the production job to use `DataflowPipelineRunner` instead
of `BlockingDataFlowPipelineRunner`, and set up the test-mode pipeline that
runs on the dataflow service so that there's now an option to use either.  As
far as I can tell, the only difference is whether your local process exits
after submitting the job or continues to run and monitor the logs, but the
former seemed more appropriate for production, so I went with it.

Test Plan:
- delete the test output table `logs_streaming_test.test_logs` in bigquery
  (Note: this is in the khan-academy project, which is not where the rest of
  our bigquery data is)
- run `make run_locally`
- verify that it streams exactly 1000 rows to bigquery by running
  `SELECT COUNT(1), MAX(start_time_timestamp) FROM [logs_streaming_test.test_logs]`
- delete the table again
- run `make deploy_test`
- verify that it starts a job on the dataflow service by checking the console at
  https://console.cloud.google.com/dataflow?project=khan-academy
- verify that it streams logs to the terminal and does not terminate the process
- verify using the same query in bigquery that it is streaming logs
  continually, and it does not stop at 1000.
- cancel the job in the cloud console, Ctrl+C out in the terminal
- run `make deploy_test_nonblocking`
- verify that the process on my local machine terminates after submitting the
  job, and that the dataflow service has a job that passes the same test plan
  as for `make deploy_test`
- cancel the job in the cloud console
- run `make deploy_prod`
- now check back in bigquery in the khanacademy.org:deductive-jet-827 project
  that running the query
  `SELECT COUNT(1), MAX(start_time_timestamp) FROM [logs_streaming.last_hour]`
  returns a nonzero count with reasonably up-to-date timestamps

Reviewers: chris, amy

Reviewed By: amy

Subscribers: tom

Differential Revision: https://phabricator.khanacademy.org/D26163