Increase the number of workers to 5
When we hit a period of high traffic this morning, the log export pipeline
failed to keep up with demand.  It turns out that what happens is that the
workers pull more and more messages and then OOM, leading to them getting
farther behind, making the problem worse.

This commit increases the number of workers to 5, which should keep up with our
peak volume (for now...).  I also changed the subscription name to clear out a
backlog of several million messages.

Test Plan:
- `make deploy_prod`, verify it's processing in production and keeping up with
  the number of messages.

Auditors: amy