Fix the streaming logs dataflow job
Summary:
Today the dataflow job that streams appengine request logs to bigquery stopped
working and throwing a ton of `IllegalArgumentException`s during log parsing.
What appears to have happened is that the format in which we receive logs
changed so that the types of the fields we're attempting to copy between two
different representations of the logs no longer match up.  Specifically the
`SourceLocation` field type was no longer compatible with the
`ArrayMap<String, Object>` we were receiving.

I have no idea why this happened; in particular, there have been no recent
updates to any of the libraries involved that seem to account for this change.
However, I don't really have a whole lot of time to investigate, so this
revision just adds some code to bridge the observed and expected data types.
This was essentially just copying keys/values from one `Map` implementation to
another, with a single conversion from `String` to `Long` added.

In order to add this bridging code, I upgraded the logging service client
library to the latest v2beta1 version from the v1beta3 version.  I'm not sure
this was strictly necessary, but I can't seem to find javadoc for the old API
version any more, so I couldn't easily proceed without the upgrade, which was
straightforward anyway.

I am now very thankful that we get to write python most of the time :)

Test Plan:
- run `make run_locally`
- verify the pipeline runs to completion
- in bigquery, run:
```
SELECT MAX(start_time_timestamp)
FROM [logs_streaming_test.test_logs]
```
- verify the date is current up to the minute, indicating that the job
  successfully streamed new logs into bigquery

Reviewers: chris, amy

Reviewed By: amy

Subscribers: tom

Differential Revision: https://phabricator.khanacademy.org/D27225