Fix a number of compile errors/warnings after auto-conversion
Summary:

After auto-converting to kotlin, there were a number of compile errors, mostly
related to the fact that several of the java APIs we rely on return Objects and
we have to cast them appropriately based on the API documentation.  (This
happens a lot for things related to JSON, like log entry data.)

This commit fixes the compile errors and most of the warnings.  There are some
warnings that remain that I haven't fixed and I'm not sure it's desirable to
fix.  These consist of unchecked casts to generic types.  Because of type
erasure, we can't, e.g., do `x as? List<String>` and have it be safe.  We'd have
to do `val y = x as? List<*>` and then `y.forEach { yi -> val z = yi as? String }`.
That's pretty ugly, especially for nested generic types, and moreover, I'm not
sure we want to be silently ignoring those errors anyway because if the casts
don't work, then that means the API changed unexpectedly on us, and that's not
good.  I'll have to think on a better way to handle situations like this.

Note that I'm pretty sure this is *not* working correctly right now.  It
compiles and runs normally, but it's streaming less than the expected number of
entries to bigquery, so I actually suspect one of the casts that I made checked
is casting to the wrong type and thus dropping data.  I'll have to go back and
be more thoughtful about it later.

Test Plan:
- `make run_locally`
- verify that the pipeline executes successfully