Enable logs from `app` to be logged to bigquery.
Summary:
This set of changes enables the sending of logs from `app` to bigquery through an already existing tool `log-export`. It achieves this by transforming the log to a form that is already understood by the application, a request_log. Additionally, it also provides a framework for transforming other types of logs if desired in the future.

Note that adding the request_id field logic in LogTransformer is pending testing until we actually manage to figure out why vm isn't sending all app logs.

Test Plan:
- Ran `make run_locally`

- Will verify that request_id parsing once vm app logs emit request log.

Reviewers: csilvers, colin

Reviewed By: csilvers, colin

Differential Revision: https://phabricator.khanacademy.org/D30749