A script to create the logs_daily/logs_hourly tables from the streaming logs.
Summary:
This script runs a bigquery query over the streaming logs that
extracts out logs for one hour.  It then emits those logs to a
logs_hourly bigquery table, and updates a logs_daily bigquery table.
It does as many hours as necessary to get everything up to speed.

Add support for merging the app_logs and request_logs from the vm module.

This makes our sql query like 100x more complicated, and also requires
introducing a temporary table.  It also makes the query much slower
and more expensive.  There is probably more we can do.

This is a WIP -- the query is still failling for us with:
> 'khanacademy.org:deductive-jet-827:bqjob_r3833898add42fa9b_00000157357f197f_1':
> 2.10306 - 2.10344: Ambiguous column name request_id.

I can't figure out where the ambiguity is.

I also added `--batch` to all the queries, since these don't need to
be on an interactive timescale.

Get the query working to merge applogs and requestlogs for managed vm's!

Boy this was complicated.  A lot of things got in the way, in the end
Sam and I had to use standard sql (not "legacy sql"), and hours and
hours of trial and error.  The end result makes sense though.

I also switched to using utcnow() everywhere, to make it clearer we're
in utc-only land.

Test Plan:
Ran
   ./create_timestamped_tables.py
and saw it create `logs_new.requestlogs_20160916_00` up through
`logs_new.requestlogs_20160916_17`, and also
`logs_new.requestlogs_20160916`.

I then manually compared logs_new.requestlogs_20160916_00 with
logs_hourly.requestlogs_20160916_00.  TODO a bit more testing before
landing.

Ran
   ./create_timestamped_tables.py

Ran the script, and visited the results in bigquery and eyeballed
them.  It will be hard to tell for sure until we expose the KALOG
lines in vm modules, which we don't right now.

Reviewers: colin, samantha

Reviewed By: colin, samantha

Differential Revision: https://phabricator.khanacademy.org/D30787