Use elog_country instead of elog_url_route for our set-per-request field.
It is unclear to me whether the url-route can sometimes be None for a
request, which would mean we couldn't tell whether this is really a
KALOG line or not.

Looking at the fields that are set on every request, elog_country is
the one that is most obvious it can never be None, since it is emitted
like this:
        log_event("country", ip_util.remote_country().upper())
(`None.upper()` fails.)  So I'll go with it.

Auditors: colin, samantha

Test plan:
None