Set up logging to log to stdout unless there's an error.
Summary:
The reason we care is this is run from cron, which will send stdout to
a logfile and stderr out as mail.

I also log timestamps, which will help a lot with debugging.

Test Plan:
I ran
   ./create_timestamped_tables.py -n > /dev/null
which produced no output, whereas before this change it will still
show the log messages.

Reviewers: samantha, colin

Differential Revision: https://phabricator.khanacademy.org/D30930