Add another check to make sure the logs are sane.
Summary:
We were seeing some problems, in our QA testing, where we compared the
hourly logs created by this script with ones created by a separate
process (the logs_to_bigquery mapreduce).  We noticed that this log
could be missing hundreds of thousands of lines.  They all -- almost
all -- were from the last 5 minutes of the log.  (This log had plenty
of loglines in the last 5 minutes, just not all of them.)  We don't
understand really how this could happen, but we can test against it
and fail.  In our experience, trying the query again an hour later
typically fixes it.

Test Plan: Mostly fingers crossed

Reviewers: samantha, colin

Reviewed By: samantha, colin

Differential Revision: https://phabricator.khanacademy.org/D30988