Say we're up to date if there are results as of 11:59:59.
This query:
```
SELECT
  MAX(end_time) >= 1474678800,
  INTEGER(MAX(end_time))
FROM
  [khan-academy:logs_streaming.logs_all_time@1474678790000-1474679760000]
```
show the last result to be at 11:59:59, even though the streaming
range on this query goes up to 12:15, which should be long enough for
the window.  It seems like maybe weird things happen at the end of the
day (UTC).  I've seen this before as well, where the logs stop at the
59:59.  Arguably that's where you want hourly logs to stop anyway.

Auditors: samantha

Test plan:
Fingers crossed