Allow there to be 'gaps' in the logs, and fill those in.
Summary:
Before, we just had a concept of 'last hourly log that was created',
and we'd start creating new ones from there.  Now we have a concept of
'hourly logs that should exist but don't', and we create all the
missing ones.

This gives us more flexibility: if we fail in processing logs for hour
X, we can still move on to X+1.

This was motivated by my observation of the logs during the 5-minute
interval starting at 2016-09-28 13:55.  This script was failing
because that interval had way fewer loglines than expected.  This went
on for like 4 hours, and it was *still* lower.  But when I looked
again 10 hours later, it was up to what I expected.

So maybe we need to wait up to 12 hours for loglines to come in.  But
I don't want to stop all logs analysis for half a day when that
happens, so I implemented this logs-filling approach instead.

Make sure we exit when we get a signal.

Otherwise timeout (which sends SIGTERM when it's time to time out)
wouldn't actually stop the script from running!

Test Plan:
I ran
   ./create_timestamped_tables.py -n
and saw it say:
   [2016-09-28 15:32:21,232 INFO] Creating the following log-tables: logs_new.requestlogs_20160928_18, logs_new.requestlogs_20160928_19, logs_new.requestlogs_20160928_20, logs_new.requestlogs_20160928_21
   [2016-09-28 15:32:21,232 INFO] Processing logs at Wed Sep 28 18:00:00 2016 (UTC)

and proceed to create a temporary table, showing it stil runs and can
process all the logs it ought to (it's currently 22:32 UTC).

Fingers crossed

Reviewers: colin, samantha

Reviewed By: samantha

Differential Revision: https://phabricator.khanacademy.org/D31088