Don't skip _hourly_logs_seem_complete even after 12 hours.
Before, we had to move on since an incomplete hour would block
all subsequent hours.  But now that we allow gaps, we don't need to
move on.  So let's keep checking forever.  I think it's better to keep
failing than to have an hour with known-missing logs.

The one downside is this could trigger on actual valid logs -- if
traffic drops precipitiously on christmas day or whatever.  But I
think the right answer will be to tweak the percent-decline that
indicates a problem.

Auditors: samantha, colin

Test plan:
None