Add a new search_harder_for_loglines option when creating the hourly table.
Summary:
We were having problems with sometimes getting fewer loglines, for a
short time, than expected:
```
 [2016-10-14 09:37:05,925 ERROR] Error reading from streaming logs: The 5 minutes of logs_new.requestlogs_20161011_22 starting at 2016-10-11 22:40:00.000000 have 21.57% fewer loglines than the preceding 5 minutes:
{"count": "865735", "interval": "2016-10-11 22:00:00.000000"}
{"count": "863702", "interval": "2016-10-11 22:05:00.000000"}
{"count": "883419", "interval": "2016-10-11 22:10:00.000000"}
{"count": "865816", "interval": "2016-10-11 22:15:00.000000"}
{"count": "873610", "interval": "2016-10-11 22:20:00.000000"}
{"count": "864874", "interval": "2016-10-11 22:25:00.000000"}
{"count": "877468", "interval": "2016-10-11 22:30:00.000000"}
{"count": "877014", "interval": "2016-10-11 22:35:00.000000"}
{"count": "687812", "interval": "2016-10-11 22:40:00.000000"}
{"count": "713555", "interval": "2016-10-11 22:45:00.000000"}
{"count": "887789", "interval": "2016-10-11 22:50:00.000000"}
{"count": "889741", "interval": "2016-10-11 22:55:00.000000"}
```

This was happening once a day or two, and keeping us from being able to rely
on these logs.

Samantha had the idea that maybe the problem was that the loglines
were just very out of order, so they weren't being captured in our
query over the logs_streaming table.

I tested this by redoing the query for the hourly log above, but
removing the end-point on the logs-streaming decorator:
```
SELECT COUNT(1) from
[khan-academy:logs_streaming.logs_all_time@1476225000000-]
WHERE end_time >= 1476225600 and end_time < 1476225900
```
And sure enough, 20,000 new loglines showed up!

So now we try that approach whenever we see a dip in loglines.  It's
expensive, but happens less than once a day so hopefully will be ok.

While doing this, I improved the logic for detecting whether there's a
dip in the loglines, by actually looking for a dip (fall and then
rise) rather than just a decrease like I was before.  This better
matches what I do when I look at these results manually.

Test Plan:
I ran
    ./create_timestamped_tables.py
It said
   [2016-10-14 11:01:17,746 INFO] Creating the following log-tables: logs_new.requestlogs_20161010_19, logs_new.requestlogs_20161011_22, logs_new.requestlogs_20161013_15, logs_new.requestlogs_20161014_17
showing that it was starting off by trying to re-generate some old
tables that had this dip-in-loglines problem.

When processing the 161010_19 logs, it said this:
   [2016-10-14 11:41:08,336 ERROR] Error reading from streaming logs at logs_new.requestlogs_20161010_19: The logs from 2016-10-10 19:55:00.000000 - the end of the hour have 22.86% fewer loglines than expected:
as expected, but then it said:
   [2016-10-14 11:43:34,023 INFO] Creating a new hourly table: logs_new.requestlogs_20161010_19
   [...]
   [2016-10-14 11:48:57,344 INFO] Updating daily table: logs_new.requestlogs_20161010
So the retry worked!  Huzzah!

Reviewers: colin, samantha

Reviewed By: samantha

Differential Revision: https://phabricator.khanacademy.org/D31382