Tighten the definition of end-of-dip.
We were seeing this 'it's a dip' report:
```
{"count": "1689267", "interval": "2016-10-17 18:00:00.000000"}
{"count": "1674601", "interval": "2016-10-17 18:05:00.000000"}
{"count": "1720500", "interval": "2016-10-17 18:10:00.000000"}
{"count": "1702706", "interval": "2016-10-17 18:15:00.000000"}
{"count": "1727473", "interval": "2016-10-17 18:20:00.000000"}
{"count": "1693401", "interval": "2016-10-17 18:25:00.000000"}
{"count": "1587157", "interval": "2016-10-17 18:30:00.000000"}  <-- dip start
{"count": "1559936", "interval": "2016-10-17 18:35:00.000000"}
{"count": "1581052", "interval": "2016-10-17 18:40:00.000000"}
{"count": "1540297", "interval": "2016-10-17 18:45:00.000000"}
{"count": "1524932", "interval": "2016-10-17 18:50:00.000000"}
{"count": "1809656", "interval": "2016-10-17 18:55:00.000000"}  <-- dip end
```
I checked with the logs_hourly table (the source of truth), and these
numbers are exactly right.  Something weird was just going on with the
last 5 minutes, where we suddenly saw a huge influx in queries.

This doesn't look like a dip (where you'd see the value going back up
to 16xxxxx instead).  I tightened the logic a bit so it doesn't think
it's a dip either.

Clearly we need better modeling here.  What I really want is some
sort of best-fit sine curve, and then noticing that one or two points
are just way out of the best-fit curve.  (Or maybe I make n best-fit
curves, one with each subset of n-1 points, and complain if one of
those curves is very different from the others.)  But that's all a
TODO for now.

Auditors: colin, samantha

Test plan:
None