Be less aggressive when we send out email.
Summary:
We are sending out email every time the input stream is a bit behind,
but it turns out that happens like all the time.  It was causing
email-spam which made it hard to see real problems.

Now I only send out email -- by sending to stderr and letting cron to
turn that into email -- when we make a final decision to keep an
hourly log even though the data looks dogdy.  That is something that a
human probably wants to know about and look in to.

Test Plan:
I ran
   ./create_timestamped_tables.py
and it said:
    [2016-10-24 14:35:43,405 INFO] Deleting logs_new.requestlogs_20161024_10: seems incomplete:
    Error reading from streaming logs at logs_new.requestlogs_20161024_10: The logs from 2016-10-24 10:40:00.000000 - 2016-10-24 10:50:00.000000 have 5.97% fewer loglines than expected:
    {"count": "299630", "interval": "2016-10-24 10:00:00.000000"}
    {"count": "306042", "interval": "2016-10-24 10:05:00.000000"}
    {"count": "315935", "interval": "2016-10-24 10:10:00.000000"}
    {"count": "317243", "interval": "2016-10-24 10:15:00.000000"}
    {"count": "324512", "interval": "2016-10-24 10:20:00.000000"}
    {"count": "322608", "interval": "2016-10-24 10:25:00.000000"}
    {"count": "326545", "interval": "2016-10-24 10:30:00.000000"}
    {"count": "324759", "interval": "2016-10-24 10:35:00.000000"}
    {"count": "305371", "interval": "2016-10-24 10:40:00.000000"}
    {"count": "308752", "interval": "2016-10-24 10:45:00.000000"}
    {"count": "322151", "interval": "2016-10-24 10:50:00.000000"}
    {"count": "318325", "interval": "2016-10-24 10:55:00.000000"}

so it looks like the new exception-based code is working as expected.

Reviewers: colin, samantha

Reviewed By: samantha

Differential Revision: https://phabricator.khanacademy.org/D31594