Move code into dataflow/log-export in preparation for moving into webapp
Summmary:

We've started keeping our dataflow jobs in our webapp repo, so this
commit changes the directory structure in preparation for moving there.
 With luck, this will allow us to `git m` this into webapp and preserve
history.

I've also deleted the LICENSE and create_timestamped_tables.py files
that won't be making the move (as well as the dotfiles).

Test Plan:
- fingers crossed

Auditors: csilvers